backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Werkzeug==3.1.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
mysql-connector-python==9.1.0
python-dotenv==1.0.1